+++ cpskin/core/browser/pages.py
@@ -407,6 +407,14 @@ class TransmoExport(BrowserView):
         portal_workflow = api.portal.get_tool('portal_workflow')
         objects['workflow'] = portal_workflow.getDefaultChain()[0]
 
+        from plone.resource.interfaces import IResourceDirectory
+        portal_resources = getUtility(IResourceDirectory, name='persistent')
+        cpskin_resources_folder = portal_resources['cpskin']
+        if cpskin_resources_folder:
+            resources = {}
+            for res in cpskin_resources_folder.listDirectory():
+                resources[res] = cpskin_resources_folder[res].data
+            objects['resources'] = resources
         portal_catalog = api.portal.get_tool('portal_catalog')
         total_objects = len(portal_catalog({}))
         objects['total_objects'] = str(total_objects)